Feat: handleErrorResponse 유틸함수 추가 Refs: #330
