+++ src/pages/MyPage/EditProfile/EditProfile.tsx
@@ -10,6 +10,7 @@ import {
 import { useLogoutMutation } from '../../../features/auth/services/authApiSlice';
 import useImageProcessing from '../../../hooks/useImageProcessing';
 import { useRouter } from '../../../hooks/useRouter';
+import handleErrorResponse from '../../../utils/handleErrorResponse';
 
 import {
   UserProfileEditForm,
@@ -117,18 +118,6 @@ const EditProfile = () => {
       }
     } catch (error) {
       console.error(error);
-    }
-  };
-
-  const handleErrorResponse = (error: any) => {
-    if ('data' in error) {
-      const data = error.data;
-      if (data !== null && typeof data === 'object' && 'message' in data) {
-        const errorMessage = data.message;
-        alert(errorMessage);
-      }
-    } else {
-      alert('잠시 후에 다시 시도해주세요.');
     }
   };
